requirements.txt
ccxt>=4.2.0
pandas>=2.0.0
ta>=0.11.0
schedule>=1.2.0
pyyaml>=6.0
requests>=2.31.0
google-generativeai>=0.3.0
